Move Validation domain tests out of root lib-test
