+++ src/tool_runtime/tests/mod.rs
@@ -48,7 +48,6 @@ mod unified_diff;
 mod validation_events;
 mod validation_handoff;
 mod validation_identity;
-mod validation_profile;
 mod validation_summary;
 mod work_on_project;
 mod write_project_file;
