+++ src/tool_runtime/validation_profile/mod.rs
@@ -1,11 +1,3 @@
-#[cfg(test)]
-mod recipe_tests;
-
 pub(crate) use webcodex_validation::{
     validation_adapter_for_tool, ValidationAdapter, ValidationCommandOptions,
-};
-#[cfg(test)]
-pub(crate) use webcodex_validation::{
-    resolve_validation_recipe, RecipeError, RecipeId, ResolvedValidationRecipe, SemanticCheck,
-    ValidationFailureEvidence,
 };
